Implement a less aggressive mode for heuristic dating supplementation.
Dating supplementation works on the assumption that when a source states
a witness has not been created after a certain date but says nothing
about the earliest possible moment it still suggests a certain closeness
of the creation date to that terminus ante quem (and the same is valid
the other way around).

The behaviour can be configured using the new option half_interval_mode:
- always: whenever a source gives only one end of the possible dating
  interval, supplement the other interval limit
- light: when we have only sources about one interval limit but none
  about the other, take the strongest source and infer the other interval
  from that.

The length of the artificial interval can be configured by half_interval_correction.