Basic implementation of a split model.
When configured using model='split', each reference (witness,
inscription etc.) is not represented by one graph node but rather by
two: one for the start of the writing process, one for the end of it.

This commit introduces the basic model, further work is required to
adjust the rest of the program to that.