Optional topologically sorted graph.
When enabling this option (e.g., in the graphviewer), the references
are arranged according to the topological ordering from the model
(that is also used for ordering the witnesses in the edition).

Note that this works by introducing additional, invisible edges, so
you shouldn’t use the graph or dot file produced using this method for
anything else than visualization purposes.